requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.11" and python_full_version <= "3.11.2"
cfgv==3.3.1 ; python_version >= "3.11" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.11" and python_version < "4.0"
django-cors-headers==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
django-filter==23.2 ; python_version >= "3.11" and python_version < "4.0"
django-redis==5.3.0 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.3 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework-simplejwt==5.2.2 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
drf-nested-routers==0.93.4 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.12.2 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.11" and python_version < "4.0"
identify==2.5.24 ; python_version >= "3.11" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.0.0 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==3.8.0 ; python_version >= "3.11" and python_version < "4.0"
pre-commit==3.3.3 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.6 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.11" and python_version < "4.0"
redis==4.6.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==68.0.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
virtualenv==20.23.1 ; python_version >= "3.11" and python_version < "4.0"
